[MIK-77][DB] - Users Table Seeder

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -23,7 +23,7 @@ class CreateUsersTable extends Migration
             $table->string('fullname')->nullable();
             $table->tinyInteger('gender')->nullable();
             $table->date('date_of_birth')->nullable();
-            $table->tinyInteger('height')->nullable();
+            $table->smallInteger('height')->nullable();
             $table->tinyInteger('salary_id')->nullable();
             $table->tinyInteger('body_type_id')->nullable();
             $table->unsignedInteger('prefecture_id')->nullable();
